Added bio to User and reworked file structure for user interaction
